feat: Initial packit setup
This commit contains the specfile for building the official package for Fedora
with a Packit setup.

Ultimately, a unified specfile is targeted for Fedora and any other rpm-based
distributions, e.g. openEuler.

And Packit(https://packit.dev/) is a tool for maintaining specfile within
upstream source. It requires a simple config file(.packit.yaml).

Log:
Change-Id: Ied633921c55452a6b56f81d4368100ef0eda6ccf
Reviewed-on: http://gerrit.uniontech.com/c/dde-api/+/8224
Reviewed-by: <mailman@uniontech.com>
Reviewed-by: songwentai <songwentai@uniontech.com>
Tested-by: <mailman@uniontech.com>